refactor Schematron project to split out classes into separate files
