+++ src/Services/Cda/Schematron.php
@@ -57,18 +57,13 @@
 
 namespace OpenEMR\Services\Cda;
 
+use DOMDocument;
+use DOMElement;
+use DOMNode;
+use InvalidArgumentException;
 use OpenEMR\Services\Cda\SchematronHelpers as Helpers;
-
-use DOMDocument,
-    DOMElement,
-    DOMNode,
-    DOMNodeList,
-    DOMXPath;
-
-use ErrorException,
-    InvalidArgumentException,
-    RuntimeException,
-    stdClass;
+use RuntimeException;
+use stdClass;
 
 class Schematron
 {
@@ -907,121 +902,4 @@ class Schematron
 
         return $type;
     }
-}
-
-
-/**
- * Helpers for work with LibXML and DOM.
- *
- * @author  Miloslav Hůla (https://github.com/milo)
- */
-class SchematronHelpers
-{
-    /** @var array */
-    private static $handleXmlErrors = array();
-
-
-    /**
-     * Enable LibXML internal error handling.
-     *
-     * @param bool  clear existing errors
-     */
-    public static function handleXmlErrors($clear = true)
-    {
-        self::$handleXmlErrors[] = libxml_use_internal_errors(true);
-        $clear && libxml_clear_errors();
-    }
-
-
-    /**
-     * Fetch all LibXML errors.
-     *
-     * @param bool
-     * @return NULL|ErrorException  all errors chained in exceptions
-     */
-    public static function fetchXmlErrors($restoreHandling = true)
-    {
-        $e = null;
-        foreach (array_reverse(libxml_get_errors()) as $error) {
-            $e = new ErrorException(trim($error->message), $error->code, $error->level, $error->file, $error->line, $e);
-        }
-        libxml_clear_errors();
-        $restoreHandling && self::restoreErrorHandling();
-        return $e;
-    }
-
-
-    /**
-     * Restore LibXML internal error handling previously enabled by self::handleXmlErrors()
-     */
-    public static function restoreErrorHandling()
-    {
-        libxml_use_internal_errors(array_pop(self::$handleXmlErrors));
-    }
-
-
-    /**
-     * Returns value of element attribute.
-     *
-     * @param DOMElement
-     * @param string  attribute name
-     * @param mixed  default value if attribude does not exist
-     * @return mixed
-     * @throws SchematronException  when attribute does not exist and default value is not specified
-     */
-    public static function getAttribute(DOMElement $element, $name)
-    {
-        if ($element->hasAttribute($name)) {
-            return $element->getAttribute($name);
-        } elseif (count($args = func_get_args()) > 2) {
-            return $args[2];
-        }
-
-        throw new SchematronException("Missing required attribute '$name' for element <$element->nodeName> on line {$element->getLineNo()}.");
-    }
-}
-
-
-/**
- * DOMXPath envelope.
- *
- * @author  Miloslav Hůla (https://github.com/milo)
- */
-class SchematronXPath extends DOMXPath
-{
-    /**
-     * ($registerNodeNS is FALSE in opposition to DOMXPath default value)
-     */
-    public function query($expression, DOMNode $context = null, $registerNodeNS = false)
-    {
-        return parent::query($expression, $context, $registerNodeNS);
-    }
-
-
-    /**
-     * ($registerNodeNS is FALSE in opposition to DOMXPath default value)
-     */
-    public function evaluate($expression, DOMNode $context = null, $registerNodeNS = false)
-    {
-        return parent::evaluate($expression, $context, $registerNodeNS);
-    }
-
-
-    public function queryContext($expression, DOMNode $context = null, $registerNodeNS = false)
-    {
-        if (isset($expression[0]) && $expression[0] !== '.' && $expression[0] !== '/') {
-            $expression = "//$expression";
-        }
-        return $this->query($expression, $context, $registerNodeNS);
-    }
-}
-
-
-/**
- * Thrown when schematron schema source is malformed (not well-formed).
- *
- * @author  Miloslav Hůla (https://github.com/milo)
- */
-class SchematronException extends \RuntimeException
-{
 }
