+++ src/Services/Cda/SchematronException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace OpenEMR\Services\Cda;
+
+/**
+ * Thrown when schematron schema source is malformed (not well-formed).
+ *
+ * @author  Miloslav Hůla (https://github.com/milo)
+ */
+class SchematronException extends \RuntimeException
+{
+}
